+++ quantforge/__init__.py
@@ -775,6 +775,7 @@ from .music import music_pseudospectrum, music_frequencies
 from .randomized_svd import randomized_svd
 from .nmf import nmf
 from .dmd import dmd
+from .cur import cur, column_leverage_scores
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1479,7 +1480,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.64.0"
+__version__ = "5.65.0"
 
 __all__ = [
     "OptionType",
@@ -2523,6 +2524,8 @@ __all__ = [
     "randomized_svd",
     "nmf",
     "dmd",
+    "cur",
+    "column_leverage_scores",
     "Dual",
     "dual_derivative",
     "dual_gradient",